Create header files for all calcs and selectors, register in LjmetFactory
