+++ LJMet/plugins/CommonCalc.cc
@@ -1,46 +1,4 @@
-/*
-  Calculator for the most common event variables
-
-   Author: Gena Kukartsev, 2012
-
-*/
-
-
-
-#include <iostream>
-#include "FWLJMET/LJMet/interface/BaseCalc.h"
-#include "FWLJMET/LJMet/interface/LjmetFactory.h"
-#include "FWLJMET/LJMet/interface/LjmetEventContent.h"
-#include "SimDataFormats/PileupSummaryInfo/interface/PileupSummaryInfo.h" 
-
-#include "PhysicsTools/Utilities/interface/LumiReWeighting.h"
-
-
-class LjmetFactory;
-
-
-class CommonCalc : public BaseCalc{
-  
-public:
-  
-  CommonCalc(){};
-  virtual ~CommonCalc(){}
-
-  virtual int BeginJob(edm::ConsumesCollector && iC);
-  virtual int AnalyzeEvent(edm::Event const & event, BaseEventSelector * selector);
-  virtual int EndJob(){return 0;}
-
-  
-private:
-    edm::EDGetTokenT<double>                           L1prefiringToken;
-    edm::EDGetTokenT<double>                           L1prefiringToken_up;
-    edm::EDGetTokenT<double>                           L1prefiringToken_down;
-
-};
-
-
-static int reg = LjmetFactory::GetInstance()->Register(new CommonCalc(), "CommonCalc");
-
+#include "FWLJMET/LJMet/interface/CommonCalc.h"
 
 
 int CommonCalc::BeginJob(edm::ConsumesCollector && iC){
