+++ LJMet/plugins/TestCalc.cc
@@ -1,25 +1,4 @@
-#include "FWLJMET/LJMet/interface/BaseCalc.h"
-#include "FWLJMET/LJMet/interface/LjmetFactory.h"
-#include "FWLJMET/LJMet/interface/LjmetEventContent.h"
-
-class LjmetFactory;
-
-class TestCalc : public BaseCalc {
-public:
-    TestCalc();
-    virtual ~TestCalc();
-    virtual int BeginJob(edm::ConsumesCollector && iC);
-    virtual int AnalyzeEvent(edm::Event const & event, BaseEventSelector * selector);
-    virtual int EndJob(){return 0;};
-    
-private:
-	bool debug;
-
-    //edm::EDGetTokenT<pat::ElectronCollection>   electronsToken;
-
-};
-
-static int reg = LjmetFactory::GetInstance()->Register(new TestCalc(), "TestCalc");
+#include "FWLJMET/LJMet/interface/TestCalc.h"
 
 
 TestCalc::TestCalc()
